[IMP] web: support custom rec_name in calendar view
Until now, the quick creation dialog of the calendar view assumed that
the underlying model used a 'name' field for its rec_name, with a
graceful degradation if it was not the case.

This prevented quick creation of models created through customizations
(for example).

This commit adds a `rec_name` attribute on the view declaration that
allows to specify the field to use during quick creation.

Task-2185423

closes odoo/odoo#48122

Related: odoo/enterprise#11242
Signed-off-by: Damien Bouvy (dbo) <dbo@odoo.com>